% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignQual-package.R
\docType{package}
\name{AlignQual}
\alias{AlignQual}
\alias{AlignQual-package}
\title{AlignQual - SAM/BAM region alignment quality metrics}
\description{
Given a sam/bam file, report a variety of quality statistics describing the
read alignments to a region. Currently only a single contiguous region is
supported, but plan to allow for a collection of regions.
}

